+++ crates/runner-nordic/crates/bootloader/build.rs
@@ -13,6 +13,6 @@
 // limitations under the License.
 
 fn main() {
-    println!("cargo:rustc-link-arg=-Tlink.x");
-    println!("cargo:rerun-if-changed=memory.x");
+    println!("cargo::rustc-link-arg=-Tlink.x");
+    println!("cargo::rerun-if-changed=memory.x");
 }